+++ src/data/objects.ts
@@ -3,6 +3,6 @@ export const enum TilemapObject {
   PLAYER_SPAWN,
 }
 
-export const objectType = new Int32Array([1, 0, 0, 0])
+export const objectType = new Uint8Array([1, 0, 0, 0])
 export const objectX = new Float64Array([320, 528, 480, 160])
 export const objectY = new Float64Array([176, 96, 256, 272])
